+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 *
 
 ### Enhancements:
-*
+* Add callback URL in session
 
 ### Fixes:
 *